Add download links for Apple and Google Play stores
Adds quick links to the store pages for the published apps.

I was sharing and noticed that there wasn't an easy way to jump to the newest release. I put it in the top section, since I think there's a good chance non-app-developers will find this via a search and want to use the app - this way they can get to the friendlier pages without being scared off.